Convert UploadHeader to TS/functional component and fix type imports
