+++ src/ts/package.json
@@ -78,6 +78,7 @@
     "resolve": "1.18.1",
     "resolve-url-loader": "^3.1.2",
     "sass-loader": "10.0.5",
+    "semantic-ui-react": "^2.0.3",
     "semver": "7.3.2",
     "style-loader": "1.3.0",
     "terser-webpack-plugin": "4.2.3",
