+++ src/ts/src/upload/types.d.ts
@@ -1,0 +1,5 @@
+export enum Step {
+    UploadSamples = "uploadSamples",
+    UploadMetadata = "uploadMetadata",
+    Review = "review",
+}
